+++ listing.h
@@ -4,7 +4,7 @@
 #ifndef     _LISTING_H
 #define     _LISTING_H
 
-void add_listing_line (char *real_line, size_t real_line_len, const char *filename, unsigned long line_number);
+void add_listing_line (char *real_line, unsigned long real_line_len, const char *filename, unsigned long line_number);
 void adjust_listings (value_t val);
 void generate_listing (void);
 void update_listing_line (struct frag *frag);
